+++ packages/identite/src/types/organization-info.ts
@@ -1,5 +1,5 @@
 //
-import type { TrancheEffectifs } from "@gouvfr-lasuite/proconnect.entreprise/types";
+import type { TrancheEffectifs } from "@proconnect-gouv/proconnect.entreprise/types";
 import { z } from "zod";
 
 //