+++ packages/identite/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @gouvfr-lasuite/proconnect.identite
+# @proconnect-gouv/proconnect.identite
+
+## 1.0.0
+
+### Major Changes
+
+- [#1455](https://github.com/proconnect-gouv/proconnect-identite/pull/1455) [`38fffb0`](https://github.com/proconnect-gouv/proconnect-identite/commit/38fffb00ca5a5d2341a662f684d2555bbfb5ff02) Thanks [@douglasduteil](https://github.com/douglasduteil)! - 📦 Première version officiel des packages `@proconnect-gouv/*`
 
 ## 0.9.0